@@ -49,12 +49,12 @@
     "mapbox-gl": "^3.8.0",
     "mdi-react": "^8.4.0",
     "moment": "^2.29.1",
-    "react": "^17.0.2",
+    "react": "^18.0.0",
     "react-bootstrap": "^2.1.2",
     "react-compound-slider": "^3.3.1",
     "react-dnd": "^15.1.1",
     "react-dnd-touch-backend": "^15.1.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.0.0",
     "react-ga4": "^1.4.1",
     "react-multiselect-dropdown-bootstrap": "^1.1.0",
     "react-query": "^3.34.15",
@@ -87,8 +87,8 @@
     "@types/mapbox__mapbox-gl-geocoder": "^4.7.0",
     "@types/mapbox__mapbox-gl-draw": "^1.4.8",
     "@types/node": "^22.7.8",
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "^17.0.11",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
     "@types/shapefile": "^0.6.4",
     "@types/shpjs": "^3.4.7",
     "@types/uuid": "^10.0.0",
