@@ -1,14 +1,9 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
+import { createRoot } from 'react-dom/client';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-import {
-  AuthenticationResult,
-  EventMessage,
-  EventType,
-  PublicClientApplication,
-} from '@azure/msal-browser';
+import { AuthenticationResult, EventMessage, EventType, PublicClientApplication } from '@azure/msal-browser';
 import { msalConfig } from './authConfig';
 
 const msalInstance = new PublicClientApplication(msalConfig);
@@ -26,13 +21,14 @@ msalInstance.addEventCallback((event: EventMessage) => {
   }
 });
 
-ReactDOM.render(
+const container = document.getElementById('root');
+const root = createRoot(container!);
+root.render(
   <React.StrictMode>
     <BrowserRouter>
       <App msalInstance={msalInstance} />
     </BrowserRouter>
   </React.StrictMode>,
-  document.getElementById('root'),
 );
 
 // If you want to start measuring performance in your app, pass a function
